[#208] EndPoint 프로토콜 구현 및 기본 구현(확장)
