+++ SobokSobok/SobokSobok/Network/API/ScheduleEndPoint.swift
@@ -12,8 +12,8 @@ enum ScheduleEndPoint {
     case getPillList(date: String)
 }
 
-extension ScheduleEndPoint {
-    var httpMethod: HttpMethod {
+extension ScheduleEndPoint: EndPoint {
+    var method: HttpMethod {
         switch self {
         case .getMySchedule:
             return .GET
@@ -22,7 +22,7 @@ extension ScheduleEndPoint {
         }
     }
     
-    var requestBody: Data? {
+    var body: Data? {
         switch self {
         case .getMySchedule:
             return nil
@@ -39,15 +39,5 @@ extension ScheduleEndPoint {
         case .getPillList(let date):
             return "\(baseURL)/schedule/detail?date=\(date)"
         }
-    }
-    
-    func createRequest(environment: APIEnvironment) -> NetworkRequest {
-        var headers: [String: String] = [:]
-        headers["Content-Type"] = "application/json"
-        headers["accesstoken"] = environment.token
-        return NetworkRequest(url: getURL(from: environment),
-                              httpMethod: httpMethod,
-                              requestBody: requestBody,
-                              headers: headers)
     }
 }
